-Made connections between the signal pins on the CP1610 and the STIC. --Not sure why the STIC has any connection to the SST, but the docs on the SST are virtually non-existent from what I could find. --I took advantage of Func and Action instead of passing bool references to both devices. I think this makes sense. -Added reset functions for both devices.
My comparison log for INTRM is still weird because it says it is true initially (Expected) and remains as such after the first instruction (A bit odd). I think this happens because the STIC is supposed to "tick" and shift SR1 to false immediately, but the STIC tick happens after the CPU tick, and the CPU tick is where the logging happens. I need to find a better place to put this logging, and I need to implement the STIC ticking for IntelliHawk. I'm not positive how to approach the latter issue as I assume a tick means one instruction execution, and my executions happen in a loop on the CPU, which has no reference to the STIC, so I'm not sure where this fits into the equation.