Fix Version project on OS X 10.11
Mono can’t be in /usr/bin/ anymore, so it wasn’t working.
It still doesn’t work now, because we’re not using SVN anymore so it
can’t run svnversion, but I figured it would be better to check this in
for when the version tool is fixed.