Fix GBC movies showing as incorrect platform
Previously, GBC movies showed their SystemId as "GB" and were
highlighted as a platform mismatch because of it.  As the emulator
infrastucture lumps GB and GBC together everywhere, this is just a GUI
hack to change the displayed platform to "GB(C)" and not highlight.

Unfortunately, Gambatte's reported system id often differs to the game
database: for both GBC and color-enabled GB games, it reports "GB"; the
database therefore cannot be used for this without significant changes
throughout the emulator.

Such changes may be a better future fix to look into, if desired.