Expanded drag and drop
Allows dragging-and-dropping of multiple files onto a running BizHawk
window.

Currently, dragging-and-dropping multiple files of the same
classification is disabled.  Classifications are listed in the
LoadOrdering Enum, and include the order in which those types of files
should be loaded (no state before ROM, etc.)
The current ordering is ROM, state, watch, CDL file, Lua session, Lua
script, cheat file, [movie file and legacy movie file].

The next thing to do is allow dropping archives and extracting the
relevant files from it, but the current hurdle is that any such loaded
file would have to be read-only, as re-saving into the archive could be
problematic.  Furthermore, all of the existing functions (copied and
split into the Loaders region) accept filenames and not streams, which
means that much of the rest of the code needs refactoring to allow
archive loading.