Get the 63 in 1 version with another property
I'm not sure if this is faithful to the hardware, but by adding another mapper property I am able to get all 4 sets of games working.

One property sets the mode (1x[63 in 1] --or-- 3x[~20 in 1])

The other property picks between the 20 in 1 sets if the first property is in that mode.

This is the only consistent way I could find to get all 4 ROMs working. But they do all work as expected so I am confident enough in the result. It is possible the game expects something else to pick between the 20 in 1 sets, but I have no idea what it could be.