Fix power up behaviour (and Dragon Slayer IV)
Power up timing is now consistent with all know information regarding VBlank timing and initial state.

This is mostly thanks to True's console testing of Battletoads, which happens to require exacting timing to sync.

However, Dragon Slayer 4 requires writes to $2000 to be recognized well before the PPU is initialized, and even blargg states that this should not be the case. Since the game obviously is supposed to run, I have removed all write ignore code until more testing can be done.