rebuild m64p
The conflict that was in #731 got unnoticed due to appveyor lying to us (because of build scripts not bombing). I fixed it now while rebasing Wyst3r's commit. But after that it would still crash in the call to connectFunctionPointers(), which required rebuilding the dll. All that should be resolved now.