Fixed SpeedThrottle()'s (relatively) high CPU usage
SpeedThrottle() was using a large portion of computation (~10 -15% of the entire program). When the sleep time is less than zero, the while true loop tags over. For some reason this appears to be occuring for more than a single. The solution is to always sleep for at least 1 ms if throttling is needed.