tastudio: disable [Ctrl+]Home/End
those are very obscure accelerators that you hit if you miss Ins/Del while editing, and that screws up the editing flow in various ways. if this functionality is needed, it will be enabled with potential miss-hitting in mind. I can't imagine a situation where you'd need them tho.

close #1600