allow informing the core that we're bulk-accessing the memory domain
since the core runs a bunch of its own threads, we have to wait for them to be ready to access the memory. hex editor and ram search read a ton of addresses at once, and we have to wait for the core for every single read.

with this new callback we set the flag once before bulk-accessing, hawk waits once for the core threads to get ready, then reads all the addresses in one go, and lets the core threads go.

speeds up hex editor with mame, even tho its mem reading is slow in general. ram search is more complex, not touching for now.