octoshock: fix up some mess in port wiring. should fix #1787 but I'm not completely sure.
also BEWARE: note the change to UpdateInput(). This was a brain-dead mistake as far as I can tell (just look at it, it seems obvious), and it came from a commit where I said "I tested it for 20 seconds" so it was probably rushed. However, fixing this to something sensible MAY affect sync. If it's important, we could roll it back. However I think UpdateInput is more like a "latch user input" than a "tick internal counters" so it may be safe