Restore the old debug mode behavior
Now, all waterbox cores will be severely hampered in speed and certain features like ram search will be essentially unusable in debug mode, as before.  This was already implemented, but it was depending only on the debug/release mode of the waterboxhost build.  We want it to be set when either the rust or the C# is compiled to debug.

This is important to catch certain dangling pointer errors.