waterbox experiment:  defensively save a lot of stuff in case something was assuming something it shouldn't
The fnstenv & stmxcsr are needed and correct since dobie plays with fenv, differently in different threads.  The rest shouldn't be because all regs in question are either volatile through calls or specified as clobbered in C code surrounding a syscall that saves them; I just did it to try to debug stuff.

In any case, none of this helps and the dobie core is still mysteriously nonfunctional