CoreInventory tweaks
When constructing a core using ConstructorInfo.Invoke, any exceptions thrown by the core are packaged inside TargetInvocationExceptions.  This has no value to us -- the fact that reflection is used is an implementation detail, and it breaks checking specific exceptions for specific information.  Accordingly, consumers have to deal with e.InnerException checks.  Fix this up so that we only expose the actual exception thrown.

In addition, undo a bad change in 6225e7854b0f3fd0be2a61ede032018deac4c88a that made the entirety of CoreInventory nonfunctional.