Move around some error messages
See #2561.  From the point of view of romloader, this is all pretty simple:  It asks for a particular settings type.  It should either get null back (indicating there was nothing, use defaults), or an object of that type.  Providing a completely unrelated type is baloney.  So this check here is a stupid defensive check that shouldn't be needed.  MainForm cannot be trusted.