Let HawkFile recover from corrupt archives (see desc., fixes #2587)
When `HawkFile` is iterating over archive members of its root trying to bind
one, if it finds one with an invalid name, it will instead bind the root as a
non-archive. This behaviour can be overridden with `forceArchiveMember: true`,
which will result in nothing bound instead of the root being bound.