Tweak the rewinder. (#2668)
MainForm intentionally frame advances right after rewinding; this is so it can capture a new framebuffer. The old rewinder sort-of understood this, the new one does not.

Tweak the logic so the new rewinder behaves like the old one, but only when the interval is 1. When the interval is greater than 1, don't go farther back than we have to.

No particular behavior is guaranteed when nonlinear time exists (ie, rewinding through manual loadstates).

Fixes #2667.