Always initialize MMF + Implement new LUA functions
This commit adresses two things:

1. Currently, the MMF LUA functions are only available if Bizhawk is started with the "--mmf=filename" flag.
I think this is not intuitive, because we should not have to specify a filename in order to active MMF functions
as the given "filename" is only used for "mmfScreenshot" (and even for this function, the filename can be set in the LUA).
I changed this behavior so that MMF functions are always available, and "mmfScreenshot" will fail with a log if
no filename as been specified by the command line nor by the LUA script.

2. I have added several LUA functions:

- "memory.writebyterangeataddress": This is because the function "memory.writebyterange" does not match with what "memory.readbyterange" returns...
"memory.readbyterange" returns a table with the keys going from 0 to len-1, while "memory.writebyterange" returns a table with each key being the full address (not starting at 0).
Actually, "memory.writebyterange" is not only for writing a range, by any set of addresses.
On the other hand, the new function "memory.writebyterangeataddress" has a behavior similar to "memory.readbyterange" or "comm.socketServerSendBytes":
the lua table given as argument is seen as an array instead of an (address, value) dictionnary.

- "mmfWriteBytes", "mmfReadBytes": to send/read a sequence of bytes to/from a memory mapped file

- "mmfCopyToMemory", "mmfCopyFromMemory": to copy the content of a memory mapped file to a given location in the emulator memory (and the other direction).
Note that this could be achieved directly in the LUA script with the combination of "memory.readbyterange" and "mmfWriteBytes", but is is A LOT slower
(probably because of the type conversions). Also note that I personaly need these functions in order to be able to efficiently read/write to the game memory
from an external tool.