(Experimental) RetroAchievements Support
Note: Users will have to manually download the RA_Integration.dll (not provided in Assets, yet). If the dll is not present, RetroAchievements stuff will not run at all and not be visible to the user.

Some of this is a bit ugly, but most of it is contained within the RetroAchievements folder. BizInvokeUtilities was modified to support [return: MarshalAs()].

Also, RA_Integration.dll's source can be found here: https://github.com/RetroAchievements/RAIntegration. It's MIT licenced so perfect for us.