Optimize TAStudio column drawing by caching PlayerNumber for each column name
TAStudio previously called `PlayerNumber(...)` for each visible cell with the column's name, only to determine whether its player number is odd or even.

Because that function uses regex and is potentially called extremely often, this had a noticable impact on fps.

I've decided to just cache the odd/even playernumber result in a dictionary for faster access, which while not being my preferred way of handling this at least results in a decent speedup.